+++ docs/CHANGELOG.md
@@ -9,7 +9,7 @@ _Latest build from master branch._
 ### Changed
 
 - Vendor [YoutubeDLSharp](https://github.com/Bluegrams/YoutubeDLSharp)
-  ([#190](https://github.com/Tubeshade/Tubeshade/issues/190))
+  ([#213](https://github.com/Tubeshade/Tubeshade/pull/213))
 - Update .NET SDK from 10.0.100 to 10.0.101, and runtime from 10.0.0 to 10.0.1
   ([#210](https://github.com/Tubeshade/Tubeshade/pull/210))
 - Update yt-dlp from 2025.11.12 to 2025.12.08